UPSTREAM: 62114: removes job scaler, continued
