+++ pkg/kubectl/cmd/BUILD
@@ -74,6 +74,7 @@ go_library(
         "//pkg/apis/core:go_default_library",
         "//pkg/apis/core/validation:go_default_library",
         "//pkg/client/clientset_generated/internalclientset:go_default_library",
+        "//pkg/client/clientset_generated/internalclientset/typed/batch/internalversion:go_default_library",
         "//pkg/client/clientset_generated/internalclientset/typed/core/internalversion:go_default_library",
         "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/apply/parse:go_default_library",
@@ -82,6 +83,7 @@ go_library(
         "//pkg/kubectl/cmd/config:go_default_library",
         "//pkg/kubectl/cmd/resource:go_default_library",
         "//pkg/kubectl/cmd/rollout:go_default_library",
+        "//pkg/kubectl/cmd/scalejob:go_default_library",
         "//pkg/kubectl/cmd/set:go_default_library",
         "//pkg/kubectl/cmd/templates:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
@@ -286,6 +288,7 @@ filegroup(
         "//pkg/kubectl/cmd/config:all-srcs",
         "//pkg/kubectl/cmd/resource:all-srcs",
         "//pkg/kubectl/cmd/rollout:all-srcs",
+        "//pkg/kubectl/cmd/scalejob:all-srcs",
         "//pkg/kubectl/cmd/set:all-srcs",
         "//pkg/kubectl/cmd/templates:all-srcs",
         "//pkg/kubectl/cmd/testdata/edit:all-srcs",
